Move 'fetch transactions' for a new block to the memPool
